Make stdlib functions dependent on reified is checks available in JS. #KT-11346 Fixed
