+++ libraries/stdlib/src/kotlin/collections/ArraysJVM.kt
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2010-2016 JetBrains s.r.o.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 @file:kotlin.jvm.JvmMultifileClass
 @file:kotlin.jvm.JvmName("ArraysKt")
 
@@ -24,9 +40,6 @@ public inline fun <reified T> Collection<T>.toTypedArray(): Array<T> {
     val thisCollection = this as java.util.Collection<T>
     return thisCollection.toArray(arrayOfNulls<T>(thisCollection.size())) as Array<T>
 }
-
-/** Returns the array if it's not `null`, or an empty array otherwise. */
-public inline fun <reified T> Array<out T>?.orEmpty(): Array<out T> = this ?: arrayOf<T>()
 
 /** Internal unsafe construction of array based on reference array type */
 internal fun <T> arrayOfNulls(reference: Array<T>, size: Int): Array<T> {
