+++ libraries/stdlib/test/collections/ArraysTest.kt
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2010-2016 JetBrains s.r.o.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package test.collections
 
 import test.collections.behaviors.listBehavior
@@ -19,6 +35,23 @@ fun assertArrayNotSameButEquals(expected: BooleanArray, actual: BooleanArray, me
 
 
 class ArraysTest {
+
+    @test fun orEmptyNull() {
+        val x: Array<String>? = null
+        val y: Array<out String>? = null
+        val xArray = x.orEmpty()
+        val yArray = y.orEmpty()
+        expect(0) { xArray.size }
+        expect(0) { yArray.size }
+    }
+
+    @test fun orEmptyNotNull() {
+        val x: Array<String>? = arrayOf("1", "2")
+        val xArray = x.orEmpty()
+        expect(2) { xArray.size }
+        expect("1") { xArray[0] }
+        expect("2") { xArray[1] }
+    }
 
     @test fun emptyArrayLastIndex() {
         val arr1 = IntArray(0)
